refactor(server-remote-cloudflare): inline auth.ts into app.ts
