+++ packages/server-remote-cloudflare/src/app.ts
@@ -1,4 +1,5 @@
 import {
+	oauthProvider,
 	oauthProviderAuthServerMetadata,
 	oauthProviderOpenIdConfigMetadata,
 } from '@better-auth/oauth-provider';
@@ -7,12 +8,18 @@ import {
 	type SupportedProvider,
 } from '@epicenter/sync-core';
 import type { Auth } from 'better-auth';
+import { betterAuth } from 'better-auth';
+import { drizzleAdapter } from 'better-auth/adapters/drizzle';
+import { bearer } from 'better-auth/plugins/bearer';
+import { jwt } from 'better-auth/plugins/jwt';
+import { drizzle } from 'drizzle-orm/postgres-js';
 import type { MiddlewareHandler } from 'hono';
 import { cors } from 'hono/cors';
 import { createFactory } from 'hono/factory';
+import postgres from 'postgres';
 import { defineErrors, type InferErrors } from 'wellcrafted/error';
 import { handleAiChat } from './ai-chat';
-import { createAuth } from './auth';
+import * as schema from './db/schema';
 
 export { YjsRoom } from './yjs-room';
 
@@ -55,6 +62,83 @@ const AuthError = defineErrors({
 	}),
 });
 type AuthError = InferErrors<typeof AuthError>;
+
+// ---------------------------------------------------------------------------
+// Auth
+// ---------------------------------------------------------------------------
+
+const trustedClients = [
+	{
+		clientId: 'epicenter-desktop',
+		name: 'Epicenter Desktop',
+		type: 'native',
+		redirectUrls: ['tauri://localhost/auth/callback'],
+		skipConsent: true,
+		metadata: {},
+	},
+	{
+		clientId: 'epicenter-mobile',
+		name: 'Epicenter Mobile',
+		type: 'native',
+		redirectUrls: ['epicenter://auth/callback'],
+		skipConsent: true,
+		metadata: {},
+	},
+] as const;
+
+/** Creates a fresh auth instance per-request. Hyperdrive clients must not be cached across requests. */
+function createAuth(env: Cloudflare.Env): AuthWithOAuth {
+	const sql = postgres(env.HYPERDRIVE.connectionString);
+	const db = drizzle(sql, { schema });
+
+	return betterAuth({
+		basePath: '/auth',
+		emailAndPassword: { enabled: true },
+		database: drizzleAdapter(db, { provider: 'pg' }),
+		baseURL: env.BETTER_AUTH_URL,
+		secret: env.BETTER_AUTH_SECRET,
+		plugins: [
+			bearer(),
+			jwt(),
+			oauthProvider({
+				loginPage: '/sign-in',
+				consentPage: '/consent',
+				requirePKCE: true,
+				allowDynamicClientRegistration: true,
+				trustedClients: [...trustedClients],
+			}),
+		],
+		session: {
+			expiresIn: 60 * 60 * 24 * 7,
+			updateAge: 60 * 60 * 24,
+			storeSessionInDatabase: true,
+			cookieCache: {
+				enabled: true,
+				maxAge: 60 * 5,
+				strategy: 'jwe',
+			},
+		},
+		advanced: {
+			crossSubDomainCookies: {
+				enabled: true,
+				domain: 'epicenter.so',
+			},
+		},
+		trustedOrigins: [
+			'https://*.epicenter.so',
+			'https://epicenter.so',
+			'tauri://localhost',
+		],
+		secondaryStorage: {
+			get: (key: string) => env.SESSION_KV.get(key),
+			set: (key: string, value: string, ttl?: number) =>
+				env.SESSION_KV.put(key, value, {
+					expirationTtl: ttl ?? 60 * 5,
+				}),
+			delete: (key: string) => env.SESSION_KV.delete(key),
+		},
+	}) as unknown as AuthWithOAuth;
+}
 
 // ---------------------------------------------------------------------------
 // Middleware
